Update pdf downloader

+++ bills/pdf/pdf.py
@@ -14,14 +14,10 @@ def get_pdf(assembly_id, start=1, end=10):
     utils.check_dir(outdir)
 
     for json in os.listdir(indir)[start:end]:
-        urls = utils.read_json('%s/%s' % (indir, json))\
-                ['status_dict']['접수']['의안접수정보'][0]['문서'][0][1]
+        url = utils.read_json('%s/%s' % (indir, json))\
+                ['status_dict']['접수']['의안접수정보'][0]['문서'][0][1][1]
+        path = '%s/%s' % (outdir, json.replace('.json', '.pdf'))
 
-        for url in urls:
-            if url.endswith('type=0'):
-                fname = json.replace('.json', '.hwp')
-            else:
-                fname = json.replace('.json', '.pdf')
-            path = '%s/%s' % (outdir, fname)
+        if not os.path.isfile(path):
             urllib.urlretrieve(url, path)
             print 'Downloaded %s' % path
